Connect CDS resolving output to feed
The CDS had support for seeing the current state of the resolution process
and controlling it, but it wasn't hooked up to the local contest - i.e. it
would control the resolver presenter and client but not the CDS itself.
This just connects it properly so that the CDS is in sync with the
resolution and resolution events go out on the regular feeds. This change
causes judgements to go out to the public feed as they resolve.

Mostly just moves the resolver logic from the resolver service (a REST
endpoint) down into the CDS contest (so an incoming step from the presenter
triggers the CDS) and ConfiguredContest (keep track of the resolution
state).

The CDS PublicContest had a regression when we enforced the data access
policy and implemented thawing: it wasn't letting resolver judgements
out. Fixed that by having a declared resolving state.

Only judgements appear on the feed, no awards yet; undo (delete) also
does not send out events.

Part of #1270.
